==> source_code/Restaurant/강남구/6. Eat_CBF_2000.ipynb <==
import pandas as pd
import numpy as np

user_df = pd.read_csv('/content/drive/MyDrive/Colab Notebooks/data/강남구_restuarant_normalized_user_matrix.csv')
place_df = pd.read_csv('/content/drive/MyDrive/Colab Notebooks/data/강남구_restuarant_normalized_item_matrix.csv')

from google.colab import drive
drive.mount('/content/drive')

user_df

del user_df['u_id']

del place_df['p_id']


user_df.reset_index(inplace=True, drop =True)
user_df

user_1 = user_df[2000:2001].T

user_1

place_df

result = np.dot(place_df, user_1) #user_1

import sys
np.set_printoptions(threshold=sys.maxsize)
result

df = pd.DataFrame(data=result, columns=["preference"])
df

place_df = pd.read_csv('/content/drive/MyDrive/Colab Notebooks/data/강남구_restuarant_normalized_item_matrix.csv')
place_df['preference'] = result

# place_df.sort_value()

place_df.sort_values(by=['preference'], axis=0, ascending=False, inplace=True)

place_df

df.reset_index(inplace=True)


df

ex = pd.read_csv('/content/drive/MyDrive/Colab Notebooks/data/강남구_restuarant_exxxx2')


ex

del ex['Unnamed: 0']

a=pd.merge(df,ex,left_on='index',right_on ='0', how='left' )
a

del a['0']


a['cbf_score']=a['preference']*a['1']
a

a.rename(columns={'index':'p_id'}, inplace=True)

a=a[['p_id', 'cbf_score']]

a

pd.DataFrame(a).to_csv('/content/drive/MyDrive/Colab Notebooks/data/prediction/강남구_restuarant_2000_cbf.csv')
