==> source_code/Restaurant/강남구/7. Combining_predictions_2000.ipynb <==
import pandas as pd
import numpy as np

user_id = int(input('user id : '))
pd.set_option('display.float_format', lambda x: '%.6f' % x)

from google.colab import drive
drive.mount('/content/drive')

eat_info =  pd.read_csv('/content/drive/MyDrive/Colab Notebooks/data/강남구_restaurant_info.csv')
cbf = pd.read_csv('/content/drive/MyDrive/Colab Notebooks/data/prediction/강남구_restuarant_'+str(user_id)+'_cbf.csv')
ncf = pd.read_csv('/content/drive/MyDrive/Colab Notebooks/data/prediction/강남구_restuarant_ncf_scores.csv')

eat_info

 eat_info=eat_info[['p_id','name','address','category','price',]]


eat_info

cbf=cbf[['p_id','cbf_score']]
cbf

ncf

ncf=ncf[ncf['u_id']==2000]

ncf

df=pd.merge(eat_info, cbf, left_on='p_id', right_on ='p_id',how ='left')
df=pd.merge(df, ncf, left_on='p_id', right_on ='p_id',how ='left')
df

df['result'] = df['cbf_score'] * df['ncf_score']

del df['u_id']

df.sort_values('result',ascending = False).head(10)
